+++ src/main/java/org/dcsa/standards/specifications/standards/tnt/v300/model/EquipmentDetails.java
@@ -7,7 +7,7 @@ import org.dcsa.standards.specifications.standards.core.v103.model.Seal;
 import org.dcsa.standards.specifications.standards.core.v103.types.EquipmentReference;
 import org.dcsa.standards.specifications.standards.core.v103.types.IsoEquipmentCode;
 import org.dcsa.standards.specifications.standards.tnt.v300.types.EmptyIndicatorCode;
-import org.dcsa.standards.specifications.standards.tnt.v300.types.MovementTypeCode;
+import org.dcsa.standards.specifications.standards.tnt.v300.types.TransportPhase;
 
 @Schema(description = "Equipment-specific details")
 @Data
@@ -20,7 +20,7 @@ public class EquipmentDetails {
 
   @Schema() private EmptyIndicatorCode emptyIndicatorCode;
 
-  @Schema() private MovementTypeCode movementType;
+  @Schema() private TransportPhase transportPhase;
 
   @Schema() private List<Seal> seals;
 }