+++ src/main/java/org/dcsa/standards/specifications/standards/tnt/v300/model/IotDetails.java
@@ -2,6 +2,7 @@ package org.dcsa.standards.specifications.standards.tnt.v300.model;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
+import org.dcsa.standards.specifications.standards.tnt.v300.types.IotEquipmentTechnologyCode;
 import org.dcsa.standards.specifications.standards.tnt.v300.types.IotEventCode;
 
 @Schema(description = "IoT-specific details")
@@ -9,4 +10,6 @@ import org.dcsa.standards.specifications.standards.tnt.v300.types.IotEventCode;
 public class IotDetails {
 
   @Schema() private IotEventCode iotEventCode;
+
+  @Schema() private IotEquipmentTechnologyCode iotEquipmentTechnologyCode;
 }