+++ src/main/java/org/dcsa/standards/specifications/standards/tnt/v300/model/ShipmentDetails.java
@@ -31,9 +31,6 @@ public class ShipmentDetails {
 
   @Schema() private ShipmentLocationTypeCode shipmentLocationType;
 
-  @Schema(description = "The reason this event was sent")
-  private String reason;
-
   @Schema(
       description =
 """