+++ src/main/java/org/dcsa/standards/specifications/standards/tnt/v300/model/EventRouting.java
@@ -8,29 +8,19 @@ import org.dcsa.standards.specifications.standards.core.v103.model.Party;
 @Schema(
     description =
 """
-Routing of an event from the originating party, via forwarding parties, to the destination party.
+Routing of an event from the originating party (or parties), optionally via one or more forwarding parties, to the destination party (or parties).
 """)
 @Data
 public class EventRouting {
 
-  @Schema(
-      description =
-"""
-The initial party with which the event originated.
-""")
-  private Party originatingParty;
+  @Schema(description = "The party (or parties) with which the event (data) originated.")
+  private List<Party> originatingParties;
 
   @Schema(
       description =
-"""
-Ordered list of parties forwarding the event between the originating and destination party.
-""")
+          "Ordered list of parties forwarding the event between the originating and destination party or parties.")
   private List<Party> forwardingParties;
 
-  @Schema(
-      description =
-"""
-The final party intended as destination of the event.
-""")
-  private Party destinationParty;
+  @Schema(description = "The party (or parties) intended as final recipients of the event.")
+  private List<Party> destinationParties;
 }