+++ src/main/resources/standards/tnt/v300/changelog.md
@@ -4,6 +4,8 @@
 
 Added IoT event subtypes for geofence entry and exit. (SD-2744)
 
+Renamed in object `EquipmentDetails` attribute `movementType` to `transportPhase`. (SD-2775)
+
 
 ## Snapshot v3.0.0-20260213-alpha